Move structs to device.
