+++ .gitignore
@@ -1,5 +1,6 @@
 .idea/
 
 VulkanTest
+*.o
 resources/shaders/frag.spv
 resources/shaders/vert.spv
