+++ resources/shaders/shader.vert
@@ -4,6 +4,7 @@ layout(binding = 0) uniform UniformBufferObject {
     mat4 model;
     mat4 view;
     mat4 proj;
+    vec3 offset;
 } ubo;
 
 layout(location = 0) in vec3 inPosition;
@@ -14,7 +15,7 @@ layout(location = 0) out vec3 fragColor;
 layout(location = 1) out vec2 fragTexCoord;
 
 void main() {
-    gl_Position = ubo.proj * ubo.view * ubo.model * vec4(inPosition, 1.0);
+    gl_Position = ubo.proj * ubo.view * ubo.model * vec4(inPosition + ubo.offset, 1.0);
     fragColor = inColor;
     fragTexCoord = inTexCoord;
 }
