--- src/queueFamilyIndices.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include <optional>
-
-// Presentation queue family could differ from graphics queue family. (p. 75)
-struct QueueFamilyIndices {
-	std::optional<uint32_t> graphicsFamily;
-	std::optional<uint32_t> presentFamily;
-
-	bool isComplete() {
-		return graphicsFamily.has_value() && presentFamily.has_value();
-	}
-};
